optional_components/speech_api/requirements.txt
gunicorn==22.0.0
uvicorn==0.23.2
openai-whisper==20231117
ffmpeg-python==0.2.0
fastapi==0.109.1
python-multipart==0.0.9
litellm==1.40.29
aiohttp==3.10.11
sounddevice==0.5.0
soundfile==0.12.1
piper-tts==1.2.0
pytest==7.4.2
pytest-asyncio==0.23.2
